minerva-doc-validator/requirements.txt
numpy
scikit-learn
sentence-transformers
nltk
pytesseract
opencv-python
Pillow
PyPDF2
python-docx
pytesseract

